Make the implied cast explicit rather than change the behaviour.

+++ src/sbbs3/sftp.cpp
@@ -1656,7 +1656,7 @@ sftp_readdir(sftp_dirhandle_t handle, void *cb_data)
 		copy_path(cwd, pm->sftp_patt);
 		while (static_files[dd->info.rootdir.idx].sftp_patt != nullptr && fn.entries() < MAX_FILES_PER_READDIR) {
 			dd->info.rootdir.idx++;
-			if (dd->info.rootdir.idx >= static_cast<int32_t>(static_files_sz))
+			if (static_cast<size_t>(dd->info.rootdir.idx) >= static_files_sz)
 				break;
 			if (static_files[dd->info.rootdir.idx].sftp_patt == nullptr) {
 				dd->info.rootdir.idx = no_more_files;
